requirements.txt
Flask
numpy
scikit-learn
joblib
python-dotenv
google-genai